Move _get_removal_time from remove action to repo shadow
