+++ docs/THOUGHTS.adoc
@@ -17,3 +17,4 @@ As I go through a major rewrite and see things I don't like, or would like to im
 - allow keyword type parameters across the connection keyword=value
 - go through all the functions and make sure their names are more consistent
 - merge somehow the ErrorLog and Log functionalities to have one single interface
+- check all tests and make sure that they don't generate uncaptured exceptions in rdiff-backup, capture them and make messages user friendly/ier
